converting inputs to Float, TODO debug inputs becoming no.float64 (Double) when made np.float32 in data download

+++ _08_cnn/utils.py
@@ -30,9 +30,10 @@ def get_datasets(config, device):
     val = np.load(path_model_inputs / 'val.npz')
     test = np.load(path_model_inputs / 'test.npz')
 
-    x_train, y_train = torch.from_numpy(train['x'][:,:-1,:,:]), torch.from_numpy(train['y'])
-    x_val, y_val = torch.from_numpy(val['x'][:,:-1,:,:]), torch.from_numpy(val['y'])
-    x_test, y_test = torch.from_numpy(test['x'][:,:-1,:,:]), torch.from_numpy(test['y'])
+    # FIXME debug Double type from model_inputs, (converted to np.float32 in data download)
+    x_train, y_train = torch.from_numpy(train['x'][:,:-1,:,:]).float(), torch.from_numpy(train['y']).float()
+    x_val, y_val = torch.from_numpy(val['x'][:,:-1,:,:]).float(), torch.from_numpy(val['y']).float()
+    x_test, y_test = torch.from_numpy(test['x'][:,:-1,:,:]).float(), torch.from_numpy(test['y']).float()
 
     # Convert nans to zeros
     x_train, y_train = torch.nan_to_num(x_train), torch.nan_to_num(y_train)
